requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
crispy-bootstrap5==0.6
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.42.0
django-ckeditor==6.0.0
django-crispy-forms==1.14.0
django-js-asset==1.2.2
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.42
psycopg2==2.9.3
psycopg2-binary==2.9.3
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
